wsd: move remote config definitions to own home
This purely moves the member function definitions
of the remote config classes into a translation
unit (RemoteConfig.cpp), to improve compile time
of COOLWSD.cpp and structure the code better.

No code changes other than (auto) formatting.
